refactor BaseballGameView class 모듈화
