+++ src/modules/views/index.js
@@ -1,0 +1,3 @@
+import BaseballGameView from './BaseballGameView';
+
+export { BaseballGameView };
